refactor(types): extract assertion functions into dedicated assertions module
