+++ src/game/services/Rules/state-machine/performingBotSetupActionState.ts
@@ -5,7 +5,7 @@ import { BOT_ACTION_DELAY } from '../../../config'
 import { incrementPlayer } from '../../../reducers/increment-player'
 import { moveCardFromHandToField } from '../../../reducers/move-card-from-hand-to-field'
 import { MatchEvent, MatchState, MatchStateGuard } from '../../../types'
-import { assertIsNonNullable } from '../../../types/guards'
+import { assertIsNonNullable } from '../../../types/assertions'
 import { botLogic } from '../../BotLogic'
 import { lookup } from '../../Lookup'
 import { GameStateCorruptError, MatchStateCorruptError } from '../errors'
