+++ src/game/services/Rules/state-machine/reducers.ts
@@ -6,7 +6,7 @@ import {
   MatchEventPayload,
   MatchEventPayloadKey,
 } from '../../../types'
-import { assertIsNonNullable } from '../../../types/guards'
+import { assertIsNonNullable } from '../../../types/assertions'
 import { lookup } from '../../Lookup'
 
 /**
