+++ src/game/types/guards/index.test.ts
@@ -1,12 +1,11 @@
 import { describe, expect, it } from 'vitest'
 
-import { CardType, MatchState } from '../'
-import { MatchStateCorruptError } from '../../services/Rules/errors'
+import { CardType } from '../'
 import { stubShovel, stubSprinkler } from '../../../test-utils/stubs/cards'
 import { carrot, instantiate, sprinkler } from '../../cards'
 import { factory } from '../../services/Factory'
 
-import * as guards from './index'
+import * as guards from '.'
 
 // Mock data for testing
 const mockCard = {
@@ -218,76 +217,6 @@ describe('Type Guards', () => {
       expect(guards.isCard({ id: 'invalid', name: 'name', type: 'CROP' })).toBe(
         false
       )
-    })
-  })
-
-  describe('assertIsCardId', () => {
-    it('does not throw for valid card ID', () => {
-      expect(() => guards.assertIsCardId('carrot')).not.toThrow()
-    })
-
-    it('throws MatchStateCorruptError for invalid card ID', () => {
-      expect(() => guards.assertIsCardId('invalid')).toThrow(
-        MatchStateCorruptError
-      )
-    })
-  })
-
-  describe('assertIsEventCardInstance', () => {
-    it('does not throw for valid Event card', () => {
-      const eventCard = { ...mockCardInstance, type: CardType.EVENT }
-
-      expect(() =>
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-argument
-        guards.assertIsEventCardInstance(eventCard as any)
-      ).not.toThrow()
-    })
-
-    it('throws MatchStateCorruptError for non-Event card', () => {
-      expect(() => guards.assertIsEventCardInstance(mockCardInstance)).toThrow(
-        MatchStateCorruptError
-      )
-    })
-  })
-
-  describe('assertIsToolCardInstance', () => {
-    it('does not throw for valid Tool card', () => {
-      const toolCard = { ...mockCardInstance, type: CardType.TOOL }
-
-      expect(() =>
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-argument
-        guards.assertIsToolCardInstance(toolCard as any)
-      ).not.toThrow()
-    })
-
-    it('throws MatchStateCorruptError for non-Tool card', () => {
-      expect(() => guards.assertIsToolCardInstance(mockCardInstance)).toThrow(
-        MatchStateCorruptError
-      )
-    })
-  })
-
-  describe('assertStringIsMatchState', () => {
-    it('does not throw for valid MatchState string', () => {
-      expect(() =>
-        guards.assertStringIsMatchState(MatchState.PLANTING_CARD)
-      ).not.toThrow()
-    })
-
-    it('throws TypeError for invalid string', () => {
-      expect(() => guards.assertStringIsMatchState('INVALID_STATE')).toThrow(
-        TypeError
-      )
-    })
-  })
-
-  describe('assertIsPlayedCrop', () => {
-    it('does not throw for valid IPlayedCrop', () => {
-      expect(() => guards.assertIsPlayedCrop(mockPlayedCrop, 0)).not.toThrow()
-    })
-
-    it('throws TypeError for undefined', () => {
-      expect(() => guards.assertIsPlayedCrop(undefined, 0)).toThrow(TypeError)
     })
   })
