+++ src/game/types/guards/index.ts
@@ -3,8 +3,6 @@ import { StateValue, StateValueMap } from 'xstate'
 import {
   CardInstance,
   CardType,
-  EventInstance,
-  MatchState,
   ICard,
   ICrop,
   ICropPriceFluctuation,
@@ -12,17 +10,14 @@ import {
   IMatch,
   IPlayedCrop,
   IPlayer,
-  isEventCardInstance,
   isToolCardInstance,
   ITable,
   ToolInstance,
   CropInstance,
   IPlayedTool,
   IPlayedCard,
-  BotTurnActionState,
 } from '../'
 import * as cards from '../../cards'
-import { MatchStateCorruptError } from '../../services/Rules/errors'
 
 export const isNonNullObject = (
   value: unknown
@@ -211,74 +206,4 @@ export const isStateValueStateValueMap = (
   stateValue: StateValue
 ): stateValue is StateValueMap => {
   return isNonNullObject(stateValue)
-}
-
-// TODO: Move assertions to their own file
-
-export function assertIsNonNullable<T>(
-  obj: T,
-  message = `${String(obj)} is null or undefined`
-): asserts obj is NonNullable<T> {
-  if (obj === undefined || obj === null) {
-    throw new TypeError(message)
-  }
-}
-
-export function assertIsCardId(id: string): asserts id is keyof typeof cards {
-  if (!isCardId(id)) {
-    throw new MatchStateCorruptError(`${id} is not a valid card ID`)
-  }
-}
-
-export function assertIsToolCardId(
-  id: string
-): asserts id is keyof typeof cards.toolCards {
-  if (!(id in cards.toolCards)) {
-    throw new MatchStateCorruptError(`${id} is not a valid tool card ID`)
-  }
-}
-
-export function assertIsEventCardInstance(
-  card: CardInstance
-): asserts card is EventInstance {
-  if (!isEventCardInstance(card)) {
-    throw new MatchStateCorruptError(`${card.id} is not an event card`)
-  }
-}
-
-export function assertIsToolCardInstance(
-  card: CardInstance
-): asserts card is ToolInstance {
-  if (!isToolCardInstance(card)) {
-    throw new MatchStateCorruptError(`${card.id} is not a tool card`)
-  }
-}
-
-export function assertStringIsMatchState(
-  str: string
-): asserts str is MatchState {
-  if (!(str in MatchState)) {
-    throw new TypeError(`${str} is not a MatchState`)
-  }
-}
-
-export function assertStateValueIsBotTurnActionState(
-  stateValue: StateValue
-): asserts stateValue is BotTurnActionState {
-  if (isStateValueStateValueMap(stateValue)) {
-    throw new TypeError(`stateValue is not a string`)
-  }
-
-  if (!(stateValue in BotTurnActionState)) {
-    throw new TypeError(`${stateValue} is not a BotTurnActionState`)
-  }
-}
-
-export function assertIsPlayedCrop(
-  plotContents: IField['cards'][0],
-  fieldCropIdx: number
-): asserts plotContents is IPlayedCrop {
-  if (plotContents === undefined) {
-    throw new TypeError(`Field plot at position ${fieldCropIdx} is undefined`)
-  }
 }
