+++ src/ui/components/Card/useCardInteractions.ts
@@ -4,10 +4,10 @@ import { STANDARD_FIELD_SIZE } from '../../../game/config'
 import { lookup } from '../../../game/services/Lookup'
 import { CardType, MatchEvent, MatchState } from '../../../game/types'
 import {
-  assertIsNonNullable,
   isCropCardInstance,
   isPlantableCardInstance,
 } from '../../../game/types/guards'
+import { assertIsNonNullable } from '../../../game/types/assertions'
 import { useMatchRules } from '../../hooks/useMatchRules'
 import { ActorContext } from '../Match/ActorContext'
 import { ShellContext } from '../Match/ShellContext'
